% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_cal.R
\name{GSEA}
\alias{GSEA}
\title{GSEA calculation}
\usage{
GSEA(gene_list, gene_set, stats_vector)
}
\arguments{
\item{gene_list}{is a list of genes.}

\item{gene_set}{is a set of genes.}

\item{stats_vector}{a vector quantify the level of genes in the gene list.}
}
\description{
This function calculates the GSEA enrichment score.
}
