% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_cal.R
\name{iES_cal}
\alias{iES_cal}
\title{iES calculation Function}
\usage{
iES_cal(Y, GSDB)
}
\arguments{
\item{Y}{is the expression matrix, GSDB is the gene set database}
}
\description{
This function allows you to express your love of cats.
}
\examples{
iES_cal()
}
\keyword{calculatioin}
\keyword{iES}
\keyword{statistics}
