% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_surv.R
\name{iES_surv}
\alias{iES_surv}
\title{iES calculation Function}
\usage{
iES_surv(GSDB, iES_mat, cli, gs_thre = 500, samp_thre = 9, qval = T)
}
\arguments{
\item{GSDB, iES_mat}{is the gene set database and iES_mat with tumor and normal}
}
\description{
This function allows to investigate on one specific pathway.
}
\examples{
iES_surv()
}
\keyword{analysis}
\keyword{and}
\keyword{for}
\keyword{groups}
\keyword{iPath}
\keyword{normal-like}
\keyword{of}
\keyword{patients:}
\keyword{perturbed}
\keyword{survival}
\keyword{two}
