% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_plots.R
\name{iES_surv_one}
\alias{iES_surv_one}
\title{iES survival for a certain pathway}
\usage{
iES_surv_one(GSDB, iES_mat, cli, gs_str, samp_thre = 9, title = T)
}
\arguments{
\item{GSDB, iES_mat, gs_str}{is the GSDB iES_mat with tumor and normal and gs name}
}
\description{
This function allows you to express your love of cats.
}
\examples{
density_fall()

}
\keyword{and}
\keyword{densityfall}
\keyword{for}
\keyword{normal}
\keyword{plot}
\keyword{sample}
\keyword{tumor}
