% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_plots.R
\name{water_fall}
\alias{water_fall}
\title{water fall plot}
\usage{
water_fall(iES_mat, gs_str, title = T)
}
\arguments{
\item{iES_mat, gs_str}{is the iES_mat with tumor and normal and gs name}
}
\description{
This function allows you to express your love of cats.
}
\examples{
water_fall()
}
\keyword{and}
\keyword{for}
\keyword{normal}
\keyword{plot}
\keyword{sample}
\keyword{tumor}
\keyword{waterfall}
